requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
openai>=1.0.0

# Vector store
pymilvus

# Web scraping
beautifulsoup4
requests
html2text==2020.1.16

# Utilities
python-dotenv
numpy>=1.26.0
tiktoken>=0.5.0
pyyaml>=6.0
